Removed the C code as I have shifted focus to solely java
