+++ nbproject/private/private.properties
@@ -1,6 +1,7 @@
 compile.on.save=true
 do.depend=false
 do.jar=true
+file.reference.quickhull3d.jar=C:\\Users\\Peter\\Desktop\\School\\Brock University\\Year 3\\COSC 3F90 - Research Project\\CPSO\\lib\\quickhull3d.jar
 javac.debug=true
 javadoc.preview=true
 user.properties.file=C:\\Users\\Peter\\AppData\\Roaming\\NetBeans\\8.1\\build.properties
