+++ nbproject/private/private.xml
@@ -3,7 +3,17 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/src/qHull/qHullJNI_QHullDelaunay.h</file>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/src/qHull/qHullTriangulate.c</file>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/src/qHull/qdelaun.c</file>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/src/qHull/qdelaunay.pro</file>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/src/qHull/poly.h</file>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/src/qHull/qhull.h</file>
             <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/src/cpso/CPSO.java</file>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/src/qHull/geom.c</file>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/src/qHull/qdelaun_r.c</file>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/src/qHull/set.h</file>
+            <file>file:/C:/Users/Peter/Desktop/School/Brock%20University/Year%203/COSC%203F90%20-%20Research%20Project/CPSO/src/Functions/QHullJNI.java</file>
         </group>
     </open-files>
 </project-private>
