--- src/qHull/README
@@ -1,62 +1,0 @@
-Name
-  qhull -- convex hull and Delaunay triangulation
-  rbox -- generate point distributions
-
-Purpose
-  The files in this directory are part of qhull-software. The
-  directory contains two programs, qhull and rbox. qhull is a general
-  dimension convex hull program that reads a set of points from stdin,
-  and outputs the smallest convex set that contains the points to
-  stdout according to printout options.  rbox is a useful tool in
-  generating input for qhull; it generates a random array of points.
-
-Environment requirements
-  qhull and rbox run on all Unix machines. ANSI-C compiler is needed to 
-  compile the files. The software is self-contained.  qhull is copyrighted 
-  software.  Please read the file COPYING.
-
-
-Authors:
-  C. Bradford Barber                    Hannu Huhdanpaa
-  The Geometry Center                   The Geometry Center
-  University of Minnesota               University of Minnesota
-  Suite 500                             Suite 500
-  1300 South Second Street,             1300 South Second Street,
-  Minneapolis, MN 55454                 Minneapolis, MN 55454
-  barber@geom.umn.edu                   hannu@geom.umn.edu
-
-Installation:
-  - change the directory variable BINDIR if needed
-  - type: make install
-  - if you want to install manual pages:
-     - change the directory variable MANDIR if needed
-     - type: make instman
-
-Product consists of:
-  documentation:
-    README
-    COPYING
-    REGISTER   /* general information on The Geometry Center shareware policy */
-    Makefile
-    qhull.ps   /* Geometry Center technical report */
-    qhull.tables.ps
-    qhull.man   /* manual page for qhull */
-    rbox.man  /* manual page for rbox */\
-    qhull.changes  /* changes made to qhull */
-  rbox consists of:
-     rbox.c
-  qhull consists of:
-     qhull.c
-     poly.c
-     geom.c
-     set.c
-     mem.c
-     globals.c
-     io.c
-     qhull.h
-     poly.h
-     geom.h
-     set.h
-     mem.h
-     globals.h
-     io.h
