--- src/qHull/rbox.man
@@ -1,66 +1,0 @@
-./"  This is the Unix manual page for rbox, written in nroff, the standard
-./"  manual formatter for Unix systems.  To format it, type
-./"
-./"  nroff -man rbox.man
-./"
-./"  This will print a formatted copy to standard output.  If you want
-./"  to ensure that the output is plain ascii, free of any control
-./"  characters that nroff uses for underlining etc, pipe the output
-./"  through "col -b":
-./"
-./"  nroff -man rbox.man | col -b
-./"
-.TH rbox 1 "May 15 1993" "Geometry Center"
-.SH NAME
-rbox \- generate random points in a cube or a sphere about the origin
-.SH SYNOPSIS
-rbox [n] [b] [s] [l] [Dn] [Wn] [Zn [Gm]] [r]
-.SH DESCRIPTION
-.PP
-rbox generates random points according to the options given, and outputs
-the points to stdout. The points are generated in a cube, unless 's' option is
-given. The format of the output is the following: first line
-contains the dimension, second line contains the number of points, and the 
-following lines contain the points, one point per line. Points are represented
-by their coordinate values.
-.SH OPTIONS
-.TP
-n
-number of points (default 10)
-.TP
-b
-add a slightly smaller cube around cospherical points ('r' turns off)
-.TP
-s
-cospherical points randomly generated in a cube and projected to the unit sphere
-.TP
-l
-spiral distribution, available only in 3-d
-.TP
-Dn
-dimension n-d (default 3-d)
-.TP
-Wn
-restrict points to distance n of the surface of a sphere or a cube
-.TP
-Zn
-restrict points to a disk about the z+ axis and the sphere (default Z1.0). Includes the opposite pole.
-.TP
-Zn Gm
-same as Zn with an empty center (default G0.5). 
-.TP
-r
-a regular polygon about the z+ axis, and projected to the unit sphere. Includes the two poles. r Z1 G0.5 gives two rings. 
-.SH BUGS
-Report bugs to qhull_bug@geom.umn.edu, other correspondence to qhull@geom.umn.edu 
-.SH SEE ALSO
-qhull(1)
-.SH AUTHOR
-.nf
-C. Bradford Barber         
-The Geometry Center        
-University of Minnesota    
-Suite 500
-1300 South Second Street,  
-Minneapolis, MN  55454     
-.fi
